refactor(cache): introduce ResourceCacheActionsTrait for enhanced resource cache management in repositories
